idastatus/requirements.txt
arrow==0.14.2
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.2.2
future==0.17.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.4.2
psycopg2-binary
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whichcraft==0.5.2
djangorestframework==3.10.2
